offset in control panel

+++ java/core/src/main/java/com/angrysurfer/core/model/Session.java
@@ -110,7 +110,7 @@ public class Session implements Serializable, IBusListener {
     private Float tempoInBPM = Constants.DEFAULT_BPM;
     private Integer loopCount = Constants.DEFAULT_LOOP_COUNT;
     private Integer parts = Constants.DEFAULT_PART_COUNT;
-    private Double noteOffset = 0.0;
+    private Integer noteOffset = 0;
 
     @Transient
     private boolean paused = false;
